% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{collapse_terms}
\alias{collapse_terms}
\title{Collapse Set of Terms Across Or}
\usage{
collapse_terms(terms = NULL, encode = FALSE)
}
\arguments{
\item{terms}{Vector of terms}

\item{encode}{If `TRUE` encodes for URL}
}
\description{
Collapse Set of Terms Across Or
}
\examples{
collapse_terms(terms = c("trump", "biden"), encode = TRUE)
}
