% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{date_columns_to_numeric}
\alias{date_columns_to_numeric}
\title{Convert date columns to numeric}
\usage{
date_columns_to_numeric(data, column_keyword = "date",
  exclude_col_type = c("character"), return_message = T)
}
\arguments{
\item{column_keyword}{Keyword of the data frame's columns to look for that signifies dates
\code{c('date')}}

\item{exclude_col_type}{The types of column to exclude from converting to numeric
\code{c('character')}}

\item{return_message}{return a message}
}
\description{
Convert date columns to numeric
}
