% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{dictionary_country_domains}
\alias{dictionary_country_domains}
\title{GDELT Outlets by Country}
\usage{
dictionary_country_domains()
}
\value{
a \code{tibble}
}
\description{
GDELT news outlets as of May 2018
}
\examples{
dictionary_country_domains()
}
