% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{dictionary_gdelt_tv_shows}
\alias{dictionary_gdelt_tv_shows}
\title{Distinct TV Shows for GDELT API}
\usage{
dictionary_gdelt_tv_shows(date_start = Sys.Date() - 30, date_end = Sys.Date())
}
\arguments{
\item{date_start}{Start Date.  Default is `Sys.Date()-1`}

\item{date_end}{End Date.  Default is Default is `Sys.Date()`}
}
\description{
Distinct TV Shows for GDELT API
}
