% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{dictionary_gdelt_tv_stations}
\alias{dictionary_gdelt_tv_stations}
\title{GDELT TV Station Dictionary}
\usage{
dictionary_gdelt_tv_stations()
}
\value{
a \code{tibble}
}
\description{
Returns information about the television
stations monitored by GDELT
}
\examples{
dictionary_gdelt_tv_stations()
}
