% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{dictionary_geo_codebook}
\alias{dictionary_geo_codebook}
\title{GDELT Geo API code book}
\usage{
dictionary_geo_codebook(code_book = "adm")
}
\arguments{
\item{code_book}{selected codebook \itemize{
\item GKG: GKG themes
\item ADM: ADM codes
\item imagetags: Image tags
\item imageweb: imageweb tags
\item languages: language of text
\item countries: country
}}
}
\value{

}
\description{
GDELT Geo API code book
}
\examples{
dictionary_geo_codebook(code_book = 'imagetags')
}
