% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{dictionary_outlets}
\alias{dictionary_outlets}
\title{GDELT Outlets}
\usage{
dictionary_outlets()
}
\description{
Returns a dictionary of GDELT's tracked outlets
including website, logo, icon and English name
}
\examples{
dictionary_outlets()
}
