% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{dictionary_stability_locations}
\alias{dictionary_stability_locations}
\title{Code book for instability locations}
\usage{
dictionary_stability_locations(snake_names = F)
}
\arguments{
\item{snake_names}{}
}
\value{
a \code{tibble}
}
\description{
Code book for instability locations
}
