% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_entities.R
\name{dictionary_v3_entity_urls}
\alias{dictionary_v3_entity_urls}
\title{V3 Entity API URL dictionary}
\usage{
dictionary_v3_entity_urls()
}
\value{

}
\description{
Dictionary of URLs from the V3 API
updated on 15 minute intervals
}
\examples{
dictionary_v3_entity_urls()
}
