% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{.plot_ft_v2_highchart}
\alias{.plot_ft_v2_highchart}
\title{Title}
\usage{
.plot_ft_v2_highchart(data, search_mode = "TimelineVolInfo",
  include_title = TRUE)
}
\arguments{
\item{include_title}{}
}
\description{
Title
}
