% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{filter_sources}
\alias{filter_sources}
\title{Filters a gdelt data frame to specified sources}
\usage{
filter_sources(data, sources = c("netsdaily", "realdeal", "curbed",
  "law360", "dailymail", "wsj.com", "law360", "pehub", "techcrunch",
  "washingtonpost", "bloomberg", "archdaily", "espn.com", "venturebeat"))
}
\arguments{
\item{data}{gkg or vgkg data frame}

\item{sources}{vector of sources}
}
\description{
Filters a gdelt data frame to specified sources
}
