% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{ft_v2_api}
\alias{ft_v2_api}
\title{Query GDELT V2 Full Text API}
\usage{
ft_v2_api(
  terms = NA,
  quote_terms = F,
  domains = NA,
  images_face_tone = NA,
  images_num_faces = NA,
  images_ocr = NA,
  images_tag = NA,
  images_web_tag = NA,
  images_web_count = NA,
  source_countries = NA,
  source_languages = "English",
  gkg_themes = NA,
  tone = NA,
  tone_absolute_value = NA,
  use_or = FALSE,
  modes = "ArtList",
  timespans = c("24 hours"),
  dates = NA,
  maximum_records = 250,
  translate = NULL,
  timeline_smooth = 5,
  sort_by = "DateDesc",
  visualize_results = TRUE,
  trelliscope_parameters = list(path = NULL, rows = 1, columns = 2, id_columns = NULL,
    group_columns = NULL),
  nest_data = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{terms}{a vector of terms}

\item{domains}{a vector of webdomains}

\item{images_face_tone}{vector of facial tone scores}

\item{images_num_faces}{vector of face count}

\item{images_ocr}{vector of words to search for OCR'd text}

\item{images_tag}{vector of image tags from the image tag code book.
use \code{dictionary_ft_codebook(code_book = "imagetag"))} for options}

\item{images_web_tag}{vector of image tags from the image web tag code book.
use \code{dictionary_ft_codebook(code_book = "imageweb"))}}

\item{images_web_count}{numeric vector of number of times photo appeared}

\item{source_countries}{character source countries
#' see \code{dictionary_ft_codebook(code_book = "countries")} for options}

\item{source_languages}{source languages}

\item{gkg_themes}{global knowledge graph theme
#' use \code{dictionary_ft_codebook(code_book = "gkg"))} for options}

\item{tone}{numeric tone - default (NA)}

\item{tone_absolute_value}{numeric tone absolute value (default NA)}

\item{use_or}{if \code{TRUE} chains multiple items using and or statement
else chained by and}

\item{modes}{API search mode \itemize{
\item  ArtList - data frame of articles for specified terms/domains/webtags/imagewebtags and OCR'd text (default)
\item ImageCollage - data frame of images from for specified terms/domains/webtags/imagewebtags and OCR'd text
\item ImageCollageInfo - data frame of images from the article for specified terms/domains/webtags/imagewebtags and OCR'd text, includes information on age of the photo and number of uses
\item ImageCollageShare - data frame of images shared on social media from the article for specified terms/domains/webtags/imagewebtags and OCR'd text, includes information on age of the photo and number of uses
\item TimelineVol - Timeline of of article volume for specified terms/domains/webtags/imagewebtags and OCR'd text
\item TimelineVolInfo - Timeline of of article volume with article information for specified terms/domains/webtags/imagewebtags and OCR'd text
\item TimelineTone - Timeline of of article tone for specified terms/domains/webtags/imagewebtags and OCR'd text
\item TimelineLang - Timeline of of article language for specified terms/domains/webtags/imagewebtags and OCR'd text
\item TimelineSourceCountry - Timeline of of article sourcelanguage for specified terms/domains/webtags/imagewebtags and OCR'd text
\item ToneChart - histogram of binned counts by tone for specified terms/domains/webtags/imagewebtags and OCR'd text
\item WordCloudNative - word cloud of native text for specified specified terms/domains/webtags/imagewebtags and OCR'd text
\item WordCloudImageTags - word cloud of resolved imagetags for specified terms/domains/webtags/imagewebtags and OCR'd text
\item WordCloudImageWebTags - word cloud of resolved image web tags for specified terms/domains/webtags/imagewebtags and OCR'd text
}}

\item{timespans}{character vector of the time frame - no more than 52 weeks -
default is 24 hours
acceptable periods include: \itemize{
\item hours (default)
\item minutes
\item weeks
\item months
\item years
}}

\item{dates}{vector of dates in YMD HMS format, separated by \code{ - }
you can use the \code{generate_dates()} function to generate a vector of
default \code{NULL}}

\item{maximum_records}{Number between 1 and 250}

\item{translate}{if \code{TRUE} translates}

\item{timeline_smooth}{if \code{mode} is a timeline}

\item{sort_by}{sorting method \itemize{
\item DateDesc - descending by date (default)
\item DateAsc - ascending by date
\item ToneDesc - descending tone
}}

\item{visualize_results}{if \code{TRUE} returns a visualization}

\item{trelliscope_parameters}{list of parameters to pass along to trelliscope \itemize{
\item path: if not \code{NULL} the path to save the trelliscope
\item rows: rows for trelliscope
\item columns: columns for trelliscope
\item id_columns: initial columns
}}

\item{nest_data}{if \code{TRUE} returns a nested \code{tibble()}}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble} or a form of visualization
}
\description{
Queries GDELT's full text API for
user specified terms, machine learned items,
web domains and more.  Users can return raw data
or a Trelliscope of visual results.  Data is a available
in rolling three month periods.
}
\examples{
terms <- c('"Brooklyn Nets"', '"Donovan Mitchell"', 'Blackston Real Estate', '"Tom Brady"')
web_sites <- c("realdeal.com", "netsdaily.com", "wsj.com", "archdaily.com", "alphr.com")
ft_v2_api(terms = terms, domains = web_sites, timespans = "28 Weeks")
}
