% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{generate_dates}
\alias{generate_dates}
\title{Generates sequences of dates}
\usage{
generate_dates(start_date = NULL, end_date = NULL,
  time_interval = "days")
}
\arguments{
\item{start_date}{date to start
if \code{NULL} defaults to 3 months ago}

\item{end_date}{date to end
if \code{NULL} defaults to system time}

\item{time_interval}{time interval sequences
\itemize{
\item months
\item weeks
\item hours
\item minutess
}}
}
\value{
a `vector` of date times
}
\description{
Generates sequences of dates
}
\examples{
generate_dates(start_date = Sys.Date()  -365, end_date = Sys.Date(), time_interval = "weeks")
}
