% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{generate_gdelt_api_parameters_values}
\alias{generate_gdelt_api_parameters_values}
\title{Generate API Query Parameters Values}
\usage{
generate_gdelt_api_parameters_values(
  parameter = "station",
  values = NULL,
  use_exact_quote = FALSE
)
}
\arguments{
\item{parameter}{Name of the Parameter}

\item{values}{Vector of Values}

\item{use_exact_quote}{If `TRUE` uses exact quote}
}
\description{
Generate API Query Parameters Values
}
\examples{
generate_gdelt_api_parameters_values(parameter = "station", values = c("CNN", "FOXNEWS", "MSNBC", "WNBC"))
generate_gdelt_api_parameters_values(parameter = "show", values = c("Hardball With Chris Matthews"),   use_exact_quote = TRUE)
}
