% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{generate_geo_query}
\alias{generate_geo_query}
\title{Generate a GDELT GEO V2 API query}
\usage{
generate_geo_query(query_parameters = list(term = NULL, domain = NULL,
  gkg_theme = NULL, image_face_tone = NULL, image_num_faces = NULL,
  image_ocr = NULL, image_tag = NULL, image_web_tag = NULL, image_web_count
  = NULL, location_name = NULL, location_adm1 = NULL, location_country =
  NULL, near = NULL, source_country = NULL, source_language = NULL, tone =
  NULL, tone_absolute_value = NULL), mode = "adm1",
  format = "ImageHTML", timespan = NULL, max_points = NULL,
  geore = NULL, sort = NULL, browse_url = TRUE)
}
\arguments{
\item{query_parameters}{list of query parameters \itemize{
 \item term: search team
 \item domain: web domain
 \item image_face_tone: image face tone
 \item image_num_face: number of faces
 \item image_ocr: OCR to find
 \item image_tag: image tagcode
 \item image_web_tag: image web code
 \item image_web_count: image count
 \item location_name: name of the location
 \item location_adm1: adm code
 \item location_country: country
 \item near: specified distances
 \item source_country: source country
 \item source_language: source language
 \item gkg_theme: GKG theme
 \item tone: numeric tone
 \item tone_absolute_value:
}}

\item{mode}{visualization mode \itemize{
\item PointData
\item ImagePointData
\item PointHeatmap
\item ImagePointheatmap
\item PointAnimation
\item ImagePointAnimation
\item Country
\item ImageCountry
\item SourceCountry
\item ImageSourceCountry
\item ADM1
\item ImageADM1
}}

\item{format}{format of the output \itemize{
\item html: HTML
\item ImageHTML: image html
\item GeoJSON: geoJSON
\item ImageGeoJSON: geoJSON with images
\item Imagehtmlshow: shows overlay-ed pictures on map
}}

\item{timespan}{time span 15 to 1440 minutes}

\item{max_points}{maximum number of points}

\item{geore}{geore level \itemize{
\item 0: All locations
\item 1: Excludes country mentions
\item 2: Only landmarks
}}

\item{sort}{how to sort the data \itemize{
\item date: by date
\item toneDesc: Tone descending
\item toneAsc: Tone ascending
}}

\item{browse_url}{if \code{TRUE} open url in browser}
}
\description{
This function will generate a V2 GEO api query and either open
the URL in the browser for exploration or download the GEOJSON data
for users to explore in R
}
\examples{
generate_geo_query(query_parameters = list(term = "Brooklyn Nets"))
}
\references{
\href{http://blog.gdeltproject.org/gdelt-geo-2-0-api-debuts/}{GDELT GEO API}
}
