% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{generate_trelliscope_bundle}
\alias{generate_trelliscope_bundle}
\title{Generate Trelliscope Bundle}
\usage{
generate_trelliscope_bundle(include_image_panel = TRUE,
  include_timeline_info = TRUE, include_timeline_tone = FALSE,
  include_sentiment_bin = TRUE,
  wordcloud_modes = c("WordCloudImageTags", "WordCloudImageWebTags"),
  terms = NA, domains = NA, gkg_themes = NA, images_ocr = NA,
  images_tag = NA, images_web_tag = NA, images_face_tone = NA,
  images_num_faces = NA, images_web_count = NA,
  timespans = c("24 hours"), wordcloud_timespans = c("1 month"),
  dates = NA, source_countries = NA, tone = NA,
  tone_absolute_value = NA, maximum_records = 250, translate = NULL,
  timeline_smooth = 5, sort_by = "DateDesc", base_path = NULL,
  rows = 1, columns = 2, id_columns = NULL, group_columns = NULL,
  return_message = TRUE)
}
\arguments{
\item{include_image_panel}{if \code{TRUE} includes ArtList}

\item{include_timeline_info}{if \code{TRUE} includes TimelineInfo}

\item{include_timeline_tone}{if \code{TRUE} includes Timeline tone}

\item{include_sentiment_bin}{if \code{TRUE} includes sentiment bin}

\item{wordcloud_modes}{Wordcloud items \itemize{
\item NA - none
\item WordCloudNative - word cloud of native text for specified specified terms/domains/webtags/imagewebtags and OCR'd text
\item WordCloudImageTags - word cloud of resolved imagetags for specified terms/domains/webtags/imagewebtags and OCR'd text
\item WordCloudImageWebTags - word cloud of resolved image web tags for specified terms/domains/webtags/imagewebtags and OCR'd text
}}

\item{terms}{a vector of terms}

\item{domains}{a vector of webdomains}

\item{gkg_themes}{global knowledge graph theme
#' use \code{dictionary_ft_codebook(code_book = "gkg"))} for options}

\item{images_ocr}{vector of words to search for OCR'd text}

\item{images_tag}{vector of image tags from the image tag code book.
use \code{dictionary_ft_codebook(code_book = "imagetag"))} for options}

\item{images_web_tag}{vector of image tags from the image web tag code book.
use \code{dictionary_ft_codebook(code_book = "imageweb"))}}

\item{images_face_tone}{vector of facial tone scores}

\item{images_num_faces}{vector of face count}

\item{images_web_count}{numeric vector of number of times photo appeared}

\item{timespans}{character vector of the time frame - no more than 12 weeks -
default is 24 hours
acceptable periods include: \itemize{
\item hours (default)
\item minutes
\item weeks
\item months
}}

\item{wordcloud_timespans}{timespans for wordcloud}

\item{dates}{vector of dates in YMD HMS format, separated by \code{ - }
you can use the \code{generate_dates()} function to generate a vector of
default \code{NULL}}

\item{source_countries}{character source countries
#' see \code{dictionary_ft_codebook(code_book = "countries")} for options}

\item{tone}{numeric tone - default (NA)}

\item{tone_absolute_value}{numeric tone absolute value (default NA)}

\item{maximum_records}{Number between 1 and 250}

\item{timeline_smooth}{if \code{mode} is a time-line}

\item{sort_by}{sorting method \itemize{
\item DateDesc - descending by date (default)
\item DateAsc - ascending by date
\item ToneDesc - descending tone
}}

\item{base_path}{base path to save the trelliscopes}

\item{rows}{rows of}

\item{return_message}{}
}
\description{
Generates a bundle of trelliscopes for publishing to the web
}
