% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_clean_count_data}
\alias{get_clean_count_data}
\title{Gets clean count data}
\usage{
get_clean_count_data(all_counts, extra_key = NA,
  count_col = "idArticle.tone", return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\value{

}
\description{
Gets clean count data
}
