% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_clean_count_vkg_data}
\alias{get_clean_count_vkg_data}
\title{Parse XML Labels}
\usage{
get_clean_count_vkg_data(all_counts, extra_key = NA,
  count_col = "idImageLabel", return_wide = T)
}
\arguments{
\item{id_dateTime}{}
}
\description{
Parse XML Labels
}
