% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_country}
\alias{get_codes_cameo_country}
\title{Retrieves GDELT CAMEO country code book}
\usage{
get_codes_cameo_country()
}
\description{
Retrieves GDELT CAMEO country code book
}
\examples{
get_codes_cameo_country()
}
