% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_ethnic}
\alias{get_codes_cameo_ethnic}
\title{Retrieves GDELT CAMEO ethnic code book}
\usage{
get_codes_cameo_ethnic()
}
\description{
Retrieves GDELT CAMEO ethnic code book
}
\examples{
get_codes_cameo_ethnic()
}
