% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_events}
\alias{get_codes_cameo_events}
\title{Retrieves CAMEO CAMEO event code book}
\usage{
get_codes_cameo_events()
}
\description{
Retrieves CAMEO CAMEO event code book
}
\examples{
get_codes_cameo_events()
}
