% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_known_groups}
\alias{get_codes_cameo_known_groups}
\title{Retrieves GDELT CAMEO known group code book}
\usage{
get_codes_cameo_known_groups()
}
\description{
Retrieves GDELT CAMEO known group code book
}
\examples{
get_codes_cameo_known_groups()
}
