% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_religion}
\alias{get_codes_cameo_religion}
\title{Retrives GDELT CAMEO religion code book}
\usage{
get_codes_cameo_religion()
}
\description{
Retrives GDELT CAMEO religion code book
}
\examples{
get_codes_cameo_religion()
}
