% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_cameo_type}
\alias{get_codes_cameo_type}
\title{Retrieves GDELT CAMEO type code book}
\usage{
get_codes_cameo_type()
}
\description{
Retrieves GDELT CAMEO type code book
}
\examples{
get_codes_cameo_type()
}
