% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_gcam}
\alias{get_codes_gcam}
\title{Retrives most recent GDELT Global Content Analysis Measures (GCAM) code book}
\usage{
get_codes_gcam()
}
\description{
Retrives most recent GDELT Global Content Analysis Measures (GCAM) code book
}
\examples{
get_codes_gcam()
}
