% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_geo_api}
\alias{get_codes_geo_api}
\title{GDELT Geo API code book}
\usage{
get_codes_geo_api(code_book = "adm")
}
\arguments{
\item{code_book}{selected codebook \itemize{
\item GKG: GKG themes
\item ADM: ADM codes
\item imagetags: Image tags
\item imageweb: imageweb tags
\item languages: language of text
\item countries: country
}}
}
\description{
GDELT Geo API code book
}
\examples{
get_codes_geo_api(code_book = 'imagetags')
}
