% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_codes_gkg_themes}
\alias{get_codes_gkg_themes}
\title{Retrieves GKG theme code book}
\usage{
get_codes_gkg_themes()
}
\description{
Retrieves GKG theme code book
}
\examples{
get_codes_gkg_themes()
}
