% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_codes_stability_locations}
\alias{get_codes_stability_locations}
\title{Code book for instability locations}
\usage{
get_codes_stability_locations()
}
\value{
a \code{tibble}
}
\description{
Code book for instability locations
}
