% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_cv_dates}
\alias{get_data_cv_dates}
\title{Gets CV Data for stated periods}
\usage{
get_data_cv_dates(dates = c("2016-06-09", "2016-06-08"),
  include_translations = F, file_directory = "Desktop/temp_gdelt_data",
  only_most_recent = F, remove_files = T, empty_trash = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Gets CV Data for stated periods
}

