% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_ft_api_domains}
\alias{get_data_ft_api_domains}
\title{Returns GDELT free text API results for multiple webdomains}
\usage{
get_data_ft_api_domains(domains = c("washingtonpost.com", "nytimes.com"),
  use_exact_domains = F, term = NA, return_image_url = T,
  last_minutes = NA, max_rows = 1000, search_language = "English",
  source_language = "English", sort_by = "date", dedeup_results = T,
  only_english = F, return_message = T)
}
\description{
Returns GDELT free text API results for multiple webdomains
}
\examples{
get_data_ft_api_domains(domains = c('realdeal.com', 'pehub.com', 'sbnation.com', 'wsj.com', 'seekingalpha.com')) \%>\% View
}

