% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_data_ft_trending_terms}
\alias{get_data_ft_trending_terms}
\title{Gets most recent terms}
\usage{
get_data_ft_trending_terms(sort_data = T)
}
\arguments{
\item{sort_data}{}
}
\description{
Gets most recent terms
}
\examples{
get_data_ft_trending_terms()
}
