% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gdelt_period_event}
\alias{get_data_gdelt_period_event}
\title{Retreive GDELT data for a given period}
\usage{
get_data_gdelt_period_event(period = 1983,
  file_directory = "Desktop/temp_gdelt_data", is_count_file = F,
  remove_files = T, empty_trash = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Retreive GDELT data for a given period
}

