% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gdelt_period_event_totals}
\alias{get_data_gdelt_period_event_totals}
\title{Retrieves GDELT event summary by period}
\usage{
get_data_gdelt_period_event_totals(period = "yearly", by_country = T,
  return_message = T)
}
\arguments{
\item{period}{can be \code{c("yearly", "daily", "monthly")}}

\item{by_country}{is data by country
can be \code{c(TRUE, FALSE)}}

\item{return_message}{returns message}
}
\description{
Retrieves GDELT event summary by period
}
\examples{
get_data_gdelt_period_event_totals(period = 'monthly', by_country = T)
}
