% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_days_detailed}
\alias{get_data_gkg_days_detailed}
\title{Get dates detailed data from a specified table}
\usage{
get_data_gkg_days_detailed(dates = c("2016-06-01"), table_name = "gkg",
  file_directory = "Desktop/temp_gdelt_data", remove_files = T,
  empty_trash = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get dates detailed data from a specified table
}

