% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_tv}
\alias{get_data_gkg_tv}
\title{Gets TV data from a URL}
\usage{

  get_data_gkg_tv(url = "http://data.gdeltproject.org/gdeltv2_iatelevision/20160609.gkg.csv.gz",
  file_directory = "Desktop/temp_gdelt_data", remove_files = T,
  empty_trash = T, return_message = T)
}
\description{
Gets TV data from a URL
}

