% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_tv_day}
\alias{get_data_gkg_tv_day}
\title{Gets GKG TV data for given day}
\usage{
get_data_gkg_tv_day(date_data = "2016-06-01", only_most_recent = F,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Gets GKG TV data for given day
}

