% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_location_instability_api}
\alias{get_data_location_instability_api}
\title{Gets instability data for given location}
\usage{
get_data_location_instability_api(location_id = "US",
  variable_name = "instability", day_moving_average = NA,
  time_period = "daily", use_multi_locations = F, return_wide = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Gets instability data for given location
}

