% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_data_locations_instability_api}
\alias{get_data_locations_instability_api}
\title{Returns instability data for given locations}
\usage{
get_data_locations_instability_api(location_ids = c("US", "IS", "TU"),
  random_locations = NULL, variable_names = c("instability",
  "conflict", "tone", "protest", "artvolnorm"), days_moving_average = NA,
  time_periods = "daily", use_multi_locations = F, return_wide = T,
  visualize = F, nest_data = F, return_message = T)
}
\arguments{
\item{location_ids}{Specify the location IDs}

\item{random_locations}{Number of random location IDs to add}

\item{variable_names}{Specify variables they can include: \itemize{
\item \code{instability}
\item \code{conflict}
\item \code{protest}
\item \code{tone}
\item \code{artvolnorm}: Relative mentions
}}

\item{days_moving_average}{Specify day moving average, NA is unsmoothed}

\item{time_periods}{Specified time period
\itemize{
\item \code{daily}
\item \code{15min}
}}

\item{use_multi_locations}{if \code{TRUE} returns all cities in a select country}

\item{return_wide}{if \code{TRUE} returns a wide data frame}

\item{visualize}{if `TRUE` returns a ggplot2 visualization}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{return_message}{if \code{TRUE} returns a location}
}
\value{
if \code{visualize} a ggplot visualization else a \code{tibble}
}
\description{
Returns instability data for given locations
}
\examples{
get_data_locations_instability_api(location_ids = c('US', 'IS', "TU"), random_locations = NULL, variable_names = c('instability', 'conflict', 'tone', 'protest', 'artvolnorm'), visualization = T, days_moving_average = NA, time_periods = 'daily', use_multi_locations = F, return_wide = T, nest_data = F, return_message = T)
}
