% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_sentiment_ft_api}
\alias{get_data_sentiment_ft_api}
\title{Returns GDELT free text API sentiment DF for a domain or a term}
\usage{
get_data_sentiment_ft_api(term = "Clinton", domain = NA,
  last_minutes = NA, is_tone = T, tone_less_than = NA,
  tone_more_than = NA, search_language = NA, source_language = NA,
  sort_by = "date", dedeup_results = T, return_message = T)
}
\value{

}
\description{
Returns GDELT free text API sentiment DF for a domain or a term
}

