% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_sentiment_ft_api_domains}
\alias{get_data_sentiment_ft_api_domains}
\title{Returns GDELT free text API word clouds for a given domain, can be term restricted}
\usage{
get_data_sentiment_ft_api_domains(domains = c("nytimes.com",
  "washingtonpost.com"), term = NA, last_minutes = NA, is_tone = T,
  tone_less_than = NA, tone_more_than = NA, search_language = NA,
  source_language = NA, sort_by = "date", dedeup_results = T,
  return_message = T)
}
\arguments{
\item{term}{options \code{c(NA, "term_name")}}

\item{return_message}{}
}
\description{
Returns GDELT free text API word clouds for a given domain, can be term restricted
}

