% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_sentiment_ft_api_terms}
\alias{get_data_sentiment_ft_api_terms}
\title{Returns GDELT free text API sentiment for a given term, can be domain restricted}
\usage{
get_data_sentiment_ft_api_terms(terms = c("Zika",
  "\\"Golden State Warriors\\""), domain = NA, last_minutes = NA,
  is_tone = T, tone_less_than = NA, tone_more_than = NA,
  search_language = NA, source_language = NA, sort_by = "date",
  dedeup_results = T, return_message = T)
}
\arguments{
\item{domain}{\code{c(NA, "domain_name")}}

\item{return_message}{}
}
\description{
Returns GDELT free text API sentiment for a given term, can be domain restricted
}
\examples{
get_data_sentiment_ft_api_terms(terms = c("Zika", '"Golden State Warriors"')) \%>\% View
}

