% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{get_data_tv_summaries}
\alias{get_data_tv_summaries}
\title{GDELT TV Show Summary}
\usage{
get_data_tv_summaries(date_start = "2018-02-02", date_end = Sys.Date(),
  return_message = T)
}
\arguments{
\item{date_start}{Start date of data, cannot be before June 16th, 2009}

\item{date_end}{End date of data, cannot exceed current date}

\item{return_message}{if \code{TRUE} returns a messag}
}
\value{
a \code{tibble}
}
\description{
Acquires daily summary data for the television shows
that GDELT monitors
}
\examples{
get_data_tv_summaries(date_start = "2018-02-02", date_end = Sys.Date(), return_message = T)
}
