% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_vgkg_day}
\alias{get_data_vgkg_day}
\title{Get Data CV Day}
\usage{
get_data_vgkg_day(date_data = "2016-06-08", include_translations = F,
  only_most_recent = F, file_directory = "Desktop/temp_gdelt_data",
  remove_files = T, empty_trash = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Get Data CV Day
}

