% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_wordcloud_ft_api}
\alias{get_data_wordcloud_ft_api}
\title{Retrives word cloud data from GDELT free text API for a given term, domain or term and domain}
\usage{
get_data_wordcloud_ft_api(term = "\\"Brooklyn Nets\\"", domain = NA,
  last_minutes = NA, search_language = "English", tone_more_than = NA,
  tone_less_than = NA, source_language = "English", sort_by = "date",
  dedeup_results = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Retrives word cloud data from GDELT free text API for a given term, domain or term and domain
}
\examples{
get_data_wordcloud_ft_api(term = NA, domain = 'wsj.com', tone_more_than = 5)
get_data_wordcloud_ft_api(term = '"Brooklyn Nets"')
}

