% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_wordcloud_ft_api_terms}
\alias{get_data_wordcloud_ft_api_terms}
\title{Returns GDELT free text API word clouds for a given term, can be domain restricted}
\usage{
get_data_wordcloud_ft_api_terms(terms = c("\\"Donald Trump\\"",
  "\\"Hilary Clinton\\""), domain = NA, last_minutes = NA,
  search_language = "English", tone_more_than = NA, tone_less_than = NA,
  source_language = "English", sort_by = "date", dedeup_results = T,
  return_message = T)
}
\arguments{
\item{domain}{\code{c(NA, "domain_name")}}

\item{return_message}{}
}
\description{
Returns GDELT free text API word clouds for a given term, can be domain restricted
}

