% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_gdelt_codebook_ft_api}
\alias{get_gdelt_codebook_ft_api}
\title{GDELT Full Text API 2.0 Codebooks}
\usage{
get_gdelt_codebook_ft_api(code_book = "gkg",
  visualize_trelliscope = FALSE, path = NULL)
}
\arguments{
\item{code_book}{code book selection \itemize{
\item gkg: Global knowledge graph codes
\item imagetags: Image tag data store
\item imageweb: Image web data store
\item languages: Languages
\item countries: Countries with country code

}}

\item{visualize_trelliscope}{if \code{TRUE} returns an
interactive trelliscope}

\item{path}{file path where you want to save the trelliscope if \code{TRUE}}
}
\description{
GDELT Full Text API 2.0 Codebooks
}
\examples{
get_gdelt_codebook_ft_api(code_book = "gkg")
get_gdelt_codebook_ft_api(code_book = "imagetags")
get_gdelt_codebook_ft_api(code_book = "imageweb")
get_gdelt_codebook_ft_api(code_book = "languages")
}
