% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_dates}
\alias{get_mentioned_gkg_dates}
\title{Returns mentioned dates from gkg data frame}
\usage{
get_mentioned_gkg_dates(gdelt_data, filter_na = T, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns mentioned dates from gkg data frame
}

