% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_event_counts}
\alias{get_mentioned_gkg_event_counts}
\title{Returns mentioned CAMEO event count from a gkg data frame}
\usage{
get_mentioned_gkg_event_counts(gdelt_data, count_column = "counts",
  filter_na = T, return_wide = F)
}
\arguments{
\item{count_column}{options \code{c('count', 'counts'))}}

\item{return_wide}{}
}
\description{
Returns mentioned CAMEO event count from a gkg data frame
}

