% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_gcams}
\alias{get_mentioned_gkg_gcams}
\title{Returns GCAM codes from a gkg data frame}
\usage{
get_mentioned_gkg_gcams(gdelt_data, merge_gcam_codes = F, filter_na = T,
  return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns GCAM codes from a gkg data frame
}

