% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_names}
\alias{get_mentioned_gkg_names}
\title{Returns mentioned names from a GKG data frame.}
\usage{
get_mentioned_gkg_names(gdelt_data, filter_na = T, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns mentioned names from a GKG data frame.
}

