% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_numerics}
\alias{get_mentioned_gkg_numerics}
\title{Returns long or wide mentioned numerics from a GKG data frame}
\usage{
get_mentioned_gkg_numerics(gdelt_data, filter_na = T, include_char_loc = T,
  return_wide = F)
}
\description{
Returns long or wide mentioned numerics from a GKG data frame
}

