% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_organizations}
\alias{get_mentioned_gkg_organizations}
\title{Returns long or wide mentioned organizations from a GKG data frame}
\usage{
get_mentioned_gkg_organizations(gdelt_data,
  organization_column = "organizations", filter_na = T, return_wide = F)
}
\arguments{
\item{organization_column}{options \code{c('organization', 'organizations', 'organizations.count', 'organizations.char_loc', 'char_loc'))}}

\item{return_wide}{}
}
\description{
Returns long or wide mentioned organizations from a GKG data frame
}

