% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_people}
\alias{get_mentioned_gkg_people}
\title{Returns long or wide mentioned people from a GKG data frame}
\usage{
get_mentioned_gkg_people(gdelt_data, people_column = "persons",
  filter_na = T, return_wide = F)
}
\arguments{
\item{people_column}{options \code{c('person', 'persons', 'persons.count', 'persons.char_loc', 'char_loc'))}}

\item{return_wide}{}
}
\description{
Returns long or wide mentioned people from a GKG data frame
}

