% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_social_embeds}
\alias{get_mentioned_gkg_social_embeds}
\title{Returns social embed information from a gkg data frame}
\usage{
get_mentioned_gkg_social_embeds(gdelt_data,
  social_embed_column = "url.social_media_image.embeds", filter_na = T,
  return_wide = F)
}
\arguments{
\item{return_wide}{}

\item{organization_column}{options \code{c('url.social_media_image.embeds', 'images', 'url.social_media_video.embeds', 'video', 'videos'))}}
}
\description{
Returns social embed information from a gkg data frame
}

