% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_mentioned_gkg_themes}
\alias{get_mentioned_gkg_themes}
\title{Returns mentioned themes from a gkg data frame}
\usage{
get_mentioned_gkg_themes(gdelt_data, filter_na = T, theme_column = "themes",
  return_wide = F)
}
\arguments{
\item{return_wide}{}

\item{organization_column}{options \code{c('theme', 'themes', 'themes.count', 'themes.char_loc', 'char_loc'))}}
}
\description{
Returns mentioned themes from a gkg data frame
}

