% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_schema_gdelt_events}
\alias{get_schema_gdelt_events}
\title{Retrieves GDELT event database schema}
\usage{
get_schema_gdelt_events()
}
\value{

}
\description{
Retrieves GDELT event database schema
}
\examples{
get_schema_gdelt_events
}
