% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_schema_gkg_counts}
\alias{get_schema_gkg_counts}
\title{Gets gkg count schema}
\usage{
get_schema_gkg_counts()
}
\value{

}
\description{
Gets gkg count schema
}
\examples{
get_schema_gkg_counts()
}
