% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_schema_gkg_general}
\alias{get_schema_gkg_general}
\title{Gets gkg general schema}
\usage{
get_schema_gkg_general()
}
\value{

}
\description{
Gets gkg general schema
}
\examples{
get_schema_gkg_general()
}
