% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_schema_gkg_mentions}
\alias{get_schema_gkg_mentions}
\title{Gets gkg mention schema}
\usage{
get_schema_gkg_mentions()
}
\value{

}
\description{
Gets gkg mention schema
}
\examples{
get_schema_gkg_mentions()
}
