% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_gdelt_event_log}
\alias{get_urls_gdelt_event_log}
\title{Gets GDELT Event data, by year from 1979-2005, by year month 2006 - 2013, then by dat}
\usage{
get_urls_gdelt_event_log(return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Gets GDELT Event data, by year from 1979-2005, by year month 2006 - 2013, then by dat
}
\examples{
get_urls_gdelt_event_log()
}
