% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_gkg_15_minute_log}
\alias{get_urls_gkg_15_minute_log}
\title{Loads gdelt v2 Global Knowledge Graph master log data, updated every 15 minutes}
\usage{
get_urls_gkg_15_minute_log()
}
\description{
Loads gdelt v2 Global Knowledge Graph master log data, updated every 15 minutes
}
\examples{
get_urls_gkg_15_minute_log()
}
