% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_gkg_daily_summaries}
\alias{get_urls_gkg_daily_summaries}
\title{Gets Global Knowledge Graph summary files by day since April 2013}
\usage{
get_urls_gkg_daily_summaries(remove_count_files = F, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Gets Global Knowledge Graph summary files by day since April 2013
}
\examples{
get_urls_gkg_daily_summaries(remove_count_files = T)
}
