% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_gkg_most_recent_log}
\alias{get_urls_gkg_most_recent_log}
\title{Gets most recent GKG log URLs}
\usage{
get_urls_gkg_most_recent_log()
}
\description{
Gets most recent GKG log URLs
}
\examples{
get_urls_gkg_most_recent_log()
}
