% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_gkg_tv_daily_summaries}
\alias{get_urls_gkg_tv_daily_summaries}
\title{Gets Global Knowledge Graph TV Summary Files}
\usage{
get_urls_gkg_tv_daily_summaries(return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Gets Global Knowledge Graph TV Summary Files
}
\examples{
get_urls_gkg_daily_summaries(remove_count_files = T)
}
