% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_urls_vgkg_most_recent}
\alias{get_urls_vgkg_most_recent}
\title{Gets most recent CV log URLs}
\usage{
get_urls_vgkg_most_recent()
}
\description{
Gets most recent CV log URLs
}
\examples{
get_urls_gkg_most_recent_log()
}
