% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R, R/gdelt_free_text_api.R
\name{load_needed_packages}
\alias{load_needed_packages}
\title{Loads needed packages}
\usage{
load_needed_packages(required_packages = c("dplyr"))

load_needed_packages(required_packages = c("dplyr"))
}
\arguments{
\item{required_packages}{}

\item{required_packages}{}
}
\value{

}
\description{
Loads needed packages

Loades Packages
}

