% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_cv_faces}
\alias{parse_cv_faces}
\title{Parses Cloud Vision Faces}
\usage{
parse_cv_faces(gdelt_data, filter_na = T, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Cloud Vision Faces
}

