% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_cv_landmarks}
\alias{parse_cv_landmarks}
\title{Parses Cloud Vision Landmark}
\usage{
parse_cv_landmarks(gdelt_data, filter_na = T, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Cloud Vision Landmark
}

