% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_article_tone}
\alias{parse_gkg_mentioned_article_tone}
\title{Returns article tones from a gkg data frame}
\usage{
parse_gkg_mentioned_article_tone(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns article tones from a gkg data frame
}
