% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_dates}
\alias{parse_gkg_mentioned_dates}
\title{Returns mentioned dates from gkg data frame}
\usage{
parse_gkg_mentioned_dates(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns mentioned dates from gkg data frame
}
