% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_gcams}
\alias{parse_gkg_mentioned_gcams}
\title{Returns GCAM codes from a gkg data frame}
\usage{
parse_gkg_mentioned_gcams(gdelt_data, merge_gcam_codes = F,
  filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns GCAM codes from a gkg data frame
}
