% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_locations}
\alias{parse_gkg_mentioned_locations}
\title{Returns mentioned locations from a gkg data frame}
\usage{
parse_gkg_mentioned_locations(gdelt_data, location_column = "locations",
  isCharLoc = F, filter_na = T, return_wide = T)
}
\arguments{
\item{location_column}{options \code{c('location', 'locations', 'locationsCharLoc', 'locationCharLoc', 'charLoc'))}}

\item{return_wide}{}
}
\description{
Returns mentioned locations from a gkg data frame
}
