% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_names}
\alias{parse_gkg_mentioned_names}
\title{Returns mentioned names from a GKG data frame.}
\usage{
parse_gkg_mentioned_names(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Returns mentioned names from a GKG data frame.
}
