% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_numerics}
\alias{parse_gkg_mentioned_numerics}
\title{Returns long or wide mentioned numerics from a GKG data frame}
\usage{
parse_gkg_mentioned_numerics(gdelt_data, filter_na = T,
  include_char_loc = T, return_wide = T)
}
\description{
Returns long or wide mentioned numerics from a GKG data frame
}
