% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_people}
\alias{parse_gkg_mentioned_people}
\title{Returns long or wide mentioned people from a GKG data frame}
\usage{
parse_gkg_mentioned_people(gdelt_data, people_column = "personsCharLoc",
  filter_na = T, return_wide = T)
}
\arguments{
\item{people_column}{options \code{c('person', 'persons', 'persons.count', 'personsCharLoc', 'charLoc'))}}

\item{return_wide}{}
}
\description{
Returns long or wide mentioned people from a GKG data frame
}
