% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_mentioned_themes}
\alias{parse_gkg_mentioned_themes}
\title{Returns mentioned themes from a gkg data frame}
\usage{
parse_gkg_mentioned_themes(gdelt_data, filter_na = T,
  theme_column = "themes", return_wide = T)
}
\arguments{
\item{organization_column}{options \code{c('theme', 'themes', 'countThemes', 'themesCharLoc', 'charLoc'))}}
}
\description{
Returns mentioned themes from a gkg data frame
}
