% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_gkg_xml_extras}
\alias{parse_gkg_xml_extras}
\title{Parses XML Extras}
\usage{
parse_gkg_xml_extras(gdelt_data, filter_na = T, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses XML Extras
}

