% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{parse_source}
\alias{parse_source}
\title{Parses Source}
\usage{

  parse_source(source = "netsdaily.com - writedate('06/02/2016 12:00 UTC'); (English / United States)")
}
\arguments{
\item{source}{}
}
\value{

}
\description{
Parses Source
}

