% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{parse_summary_inventory_data_urls}
\alias{parse_summary_inventory_data_urls}
\title{Parsse Inventory URLs}
\usage{
parse_summary_inventory_data_urls(
  urls = "http://data.gdeltproject.org/gdeltv3/iatv/inventory/20240805.inventory.csv",
  snake_names = FALSE,
  return_message = T
)
}
\arguments{
\item{urls}{vector of GDELT v3 URLSs}

\item{snake_names}{if `TRUE` snakes names}

\item{return_message}{if `TRUE` returns message}
}
\description{
Parsse Inventory URLs
}
\examples{
parse_summary_inventory_data_urls("http://data.gdeltproject.org/gdeltv3/iatv/inventory/20240805.inventory.csv")
}
