% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{parse_v2_tv_api_urls}
\alias{parse_v2_tv_api_urls}
\title{Parse V2 TV API URL}
\usage{
parse_v2_tv_api_urls(urls = NULL, nest_data = FALSE, return_message = T)
}
\arguments{
\item{urls}{Vector of URLS}

\item{nest_data}{If `TRUE` nests data}

\item{return_message}{if `TRUE` returns message}
}
\description{
Parse V2 TV API URL
}
