% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_entities.R
\name{parse_v3_entity_api_urls}
\alias{parse_v3_entity_api_urls}
\title{Parse V3 Entity API URLs}
\usage{

  parse_v3_entity_api_urls(urls = c("http://data.gdeltproject.org/gdeltv3/geg_gcnlapi/20190214011500.geg-gcnlapi.json.gz",
  "http://data.gdeltproject.org/gdeltv3/geg_gcnlapi/20190130204500.geg-gcnlapi.json.gz",
  "http://data.gdeltproject.org/gdeltv3/geg_gcnlapi/20190912134500.geg-gcnlapi.json.gz"),
  return_message = TRUE)
}
\arguments{
\item{urls}{vector of URLs}

\item{return_message}{if \code{TRUE} returns message}
}
\value{
\code{tibble}
}
\description{
Parse vector of V3 entity API urls
}
