% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_links.R
\name{parse_v3_social_urls}
\alias{parse_v3_social_urls}
\title{Parse V3 Social Link URLs}
\usage{
parse_v3_social_urls(urls = NULL, return_message = T)
}
\arguments{
\item{urls}{vector of social link urls}

\item{return_message}{if \code{TRUE} rturns message}
}
\value{

}
\description{
Parse V3 Social Link URLs
}
