% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_vgkg_faces}
\alias{parse_vgkg_faces}
\title{Parses Cloud Vision Faces}
\usage{
parse_vgkg_faces(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Cloud Vision Faces
}
