% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_vgkg_languages}
\alias{parse_vgkg_languages}
\title{Parses Cloud Vision Languages}
\usage{
parse_vgkg_languages(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Cloud Vision Languages
}
