% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_vgkg_ocr}
\alias{parse_vgkg_ocr}
\title{Parses Cloud Vision OCR}
\usage{
parse_vgkg_ocr(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Cloud Vision OCR
}
