% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_vgkg_safe_search}
\alias{parse_vgkg_safe_search}
\title{Parses Safe Search}
\usage{
parse_vgkg_safe_search(gdelt_data, filter_na = T, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Parses Safe Search
}
