% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_xml_extras}
\alias{parse_xml_extras}
\title{Parse XML Labels}
\usage{
parse_xml_extras(data,
  id_gkg = "20160601010000-BLOOMBERG_20160601_010000_Trending_Business")
}
\arguments{
\item{id_dateTime}{}
}
\description{
Parse XML Labels
}

