% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_xml_safe_search}
\alias{parse_xml_safe_search}
\title{Parses Safe Search}
\usage{
parse_xml_safe_search(data, id_vgkg = 1)
}
\arguments{
\item{id_vgkg}{}
}
\value{

}
\description{
Parses Safe Search
}
