% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{plot_trelliscopes}
\alias{plot_trelliscopes}
\title{Plot list of trelliscope}
\usage{
plot_trelliscopes(data, group_columns = NULL,
  trelliscope_parameters = list(path = NULL, rows = 1, columns = 2))
}
\arguments{
\item{data}{a data frame}

\item{group_columns}{columns for grouped data frame}

\item{trelliscope_parameters}{list of parameters to pass along to trelliscope \itemize{
\item path: if not \code{NULL} the path to save the trelliscope
\item rows: rows for trelliscope
\item columns: columns for trelliscope
\item id_columns: initial columns
}}
}
\description{
Plot list of trelliscope
}
