% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{remove_full_na_column}
\alias{remove_full_na_column}
\title{Removes full NA columns from specified data frame}
\usage{
remove_full_na_column(data, col_name = c("codeLocation"))
}
\arguments{
\item{data}{a data frame}

\item{col_name}{specified column}
}
\description{
Removes full NA columns from specified data frame
}

