% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{v2_tv_api}
\alias{v2_tv_api}
\title{GDELT V2 TV API Data and URLS}
\usage{
v2_tv_api(
  terms = NULL,
  use_exact_term = FALSE,
  parse_data = TRUE,
  return_message = TRUE,
  networks = NULL,
  stations = NULL,
  contexts = NULL,
  markets = NULL,
  shows = NULL,
  modes = "ClipGallery",
  formats = "json",
  use_combined_stations = FALSE,
  data_normalization_parameter = "percent",
  date_resolution = NULL,
  include_last_24_hours = FALSE,
  maximum_records = 3000,
  sort_variable = "DateDesc",
  timeline_smooth = NULL,
  start_date = NULL,
  end_date = NULL,
  timespan = NULL,
  timezone_adjust = NULL,
  time_zoom = NULL
)
}
\arguments{
\item{terms}{Vector of terms}

\item{use_exact_term}{If `TRUE` uses exact term}

\item{parse_data}{If `TRUE` parses data}

\item{return_message}{If `TRUE` returns a message}

\item{networks}{Vector of networks}

\item{stations}{Vector of stations}

\item{contexts}{Vector of contexts}

\item{markets}{Vector of markets}

\item{shows}{Vector of shows}

\item{modes}{\itemize{
\item `ClipGallery` (default) This displays up to the top 50 most relevant clips matching your search. Each returned clip includes the name of the source show and station, the time the clip aired, a thumbnail, the actual text of the snippet and a link to view the full one minute clip on the Internet Archive's website. This allows you to see what kinds of clips are matching and view the full clip to gain context on your search results. In HTML output, this mode displays a "high design" visual layout suitable for creating magazine-style collages of matching coverage. When embedded as an iframe, the API uses the same postMessage resize model as the DOC 2.0 API.
\item `ShowChart` This determines what percent of your search results were from each specific television show and displays the top 10 shows
\item `StationChart` This compares how many results your search generates from each of the selected stations over the selected time period, allowing you to assess the relative attention each is paying to the topic. Using the DATANORM parameter you can control whether this reports results as raw clip counts or as normalized percentages of all coverage (the most robust way of comparing stations). Note that in HTML mode, you can use the button at the top right of the display to save it as a static image or save its underlying data.
\item `TimelineVol` This tracks how many results your search generates by day/hour over the selected time period, allowing you to assess the relative attention each is paying to the topic and how that attention has varied over time. Using the DATANORM parameter you can control whether this reports results as raw clip counts or as normalized percentages of all coverage (the most robust way of comparing stations). By default, the timeline will not display the most recent 24 hours, since those results are still being generated (it can take up to 2-12 hours for a show to be processed by the Internet Archive and ready for analysis), but you can include those if needed via the LAST24 option. You can also smooth the timeline using the TIMELINESMOOTH option and combine all selected stations into a single time series using the DATACOMB option. Note that in HTML mode, you can toggle the station legend using the button that appears at the top right of the display or export the timeline as a stati
\item `TimelineVolHeatmap` Displays an hourly timeline of total airtime seconds matching the query, rapidly pinpointing hourly trends, where the X axis is days and Y axis is hours from 0 to 23. Each cell is color-coded from white (0) to dark blue (maximum value). Note that this visualization is very computationally expensive and thus may take several seconds or longer to return
\item `TimelineVolStream` Same as "TimelineVol" but displays as a streamgraph rather than a line-based timeline
\item `TimelineVolNorm` This displays the total airtime (in terms of 15 second clips) monitored from each of the stations in your query. It must be combined with a valid query, since it displays the airtime for the stations queried in the search. This mode can be used to identify brief monitoring outages or for advanced normalization, since it reports the total amount of clips monitored overall from each station in each day/hour.
\item `WordCloud` This mode returns the top words that appear most frequently in clips matching your search. It takes the 200 most relevant clips that match your search and displays a word cloud of up to the top 200 most frequent words that appeared in those clips (common stop words are automatically removed). This is a powerful way of understanding the topics and words dominating the relevant coverage and suggesting additional contextual search terms to narrow or evolve your search. Note that if there are too few matching clips for your query, the word cloud may be blank. Note that in HTML mode, you can use the options at the bottom right of the display to save it as a static image or save its underlying data
}}

\item{formats}{This controls what file format the results are displayed in. Not all formats are available for all modes.  To assist with website embedding, the CORS ACAO header for all output of the API is set to the wildcard "*", permitting universal embedding. \itemize{
\item `HTML` This is the default mode and returns a browser-based visualization or display. Some displays, such as word clouds, are static images, some, like the timeline modes, result in interactive clickable visualizations, and some result in simple HTML lists of images or articles. The specific output varies by mode, but all are intended to be displayed directly in the browser in a user-friendly intuitive display and are designed to be easily embedded in any page via an iframe.
\item `CSV` This returns the requested data in comma-delimited (CSV) format. The specific set of columns varies based on the requested output mode. Note that since some modes return multilingual content, the CSV is encoded as UTF8 and includes the UTF8 BOM to work around Microsoft Excel limitations
\item `JSON` This returns the requested data in UTF8 encoded JSON. The specific fields varies by output mode.
\item `RSS` This is only available in ClipGallery mode and returns the results in RSS 2.0 format
}}

\item{use_combined_stations}{If `TRUE` both timeline and station chart modes separate out each matching station's data to make it possible to compare the relative attention paid to a given topic by each station. Sometimes, however, the interest is in overall media attention, rather than specific per-station differences in that coverage. Setting this parameter to "combined" will collapse all matching data into a single "Combined" synthetic station and make it much easier to understand macro-level patterns}

\item{data_normalization_parameter}{Most media researchers are accustomed to working with raw result counts, such as the absolute number of 15 second clips that matched a given query. Such raw counts, while useful quick "sanity checks", are inappropriate for production analysis, especially looking over time or across stations. Raw clip counts reflect the absolute number of clips that matched a given query. However, one station may devote more of its airtime to commercials or non-news entertainment programming (which are largely excluded by the TV Explorer) and thus render such comparisons meaningless. Instead, by default the API normalizes timeline and station chart modes by dividing the number of matching clips in each time interval by the total number of all monitored clips from that station or set of stations over that time interval. For example, if searching for "trump" on CNN over the past 2 years in timeline mode, each day will be displayed as the percent of all 15 second clips monitored from CNN that day that contained the word "trump" somewhere in that 15 second interval. This transforms result counts into normalized density measurements that allow you to directly compare different periods of time or stations in a meaningful way. We strongly recommend that users use only normalized result counts (the default), but for cases where you need raw counts, you can set this parameter to.
\itemize{
\item `percent` This is the default mode and returns the results as a percentage of all monitored clips from that station or set of stations over that time interval.
\item `raw` This returns the results as raw counts of matching clips.
}}

\item{date_resolution}{By default results are returned in hourly resolution (for <7 day timespans), daily resolution for <3 year timespans and monthly resolution otherwise. You can override these settings and manually set the time resolution to any of the following values. NOTE that this will automatically adjust the STARTDATETIME and ENDDATETIME parameters to their start/stop dates/times at the given date resolution. \itemize{
\item `Hour:` This is only available for timespans of 7 or fewer days. Rounds start/stop dates to their nearest full hours.
\item `Day`: Rounds start/stop dates to their nearest full day.
\item `Week`: Rounds start/stop dates to their nearest full week.
\item `Month`: Rounds start/stop dates to their nearest full month.
\item `Year`: Rounds start/stop dates to their nearest full year.
}}

\item{include_last_24_hours}{If `TRUE` includes the most recent 24 hours of data in the timeline. By default, the timeline will not display the most recent 24 hours, since those results are still being generated (it can take up to 2-12 hours for a show to be processed by the Internet Archive and ready for analysis), but you can include those if needed via the LAST24 option.}

\item{maximum_records}{This option only applies to ClipGallery mode. By default 50 clips are displayed in HTML mode. In JSON, JSONP or CSV formats, up to 3,000 clips can be returned}

\item{sort_variable}{By default results are sorted by relevance to your query. Sometimes you may wish to sort by date or tone instead.}

\item{timeline_smooth}{This option is only available in timeline mode and performs moving window smoothing over the specified number of time steps, up to a maximum of 30. Timeline displays can sometimes capture too much of the chaotic noisy information environment that is the television landscape, resulting in jagged displays. Use this option to enable moving average smoothing up to 30 time steps to smooth the results to see the macro-level patterns. Note that since this is a moving window average, peaks will be shifted to the right, up to several days or weeks at the heaviest smoothing levels}

\item{start_date}{Start date; The earliest available date is July 2, 2009. If you do not specify an ENDDATETIME, the API will search from STARTDATETIME through the present date/time}

\item{end_date}{End Date;  If you do not specify a STARTDATETIME, the API will search from July 2, 2009 through the specified ENDDATETIMEs}

\item{timespan}{By default the TV API searches the entirety of the Internet Archive's Television News Archive's holdings, which extend back to July 2009 for some stations. You can narrow this range by using this option to specify the number of months, weeks, days or hours (minimum of 1 hour). The API then only searches airtime within the specified timespan backwards from the present time. If you would instead like to specify the precise start/end time of the search instead of an offset from the present time, you should use the STARTDATETIME/ENDDATETIME parameters
\itemize{
\item `Hours`. Specify a number followed by "h" or "hours" to provide the timespan in hours.
\item  `Days`. Specify a number followed by "d" or "days" to provide the timespan in days.
 \item `Weeks`. Specify a number followed by "w" or "weeks" to provide the timespan in weeks.
 \item `Months`. Specify a number followed by "m" or "months" to provide the timespan in months.
 \item `Years`. Specify a number followed by "y" or "years" to provide the timespan in years.

}}

\item{timezone_adjust}{By default STARTDATETIME, ENDDATETIME and all returned results are interpreted and reported in the UTC timezone. Use this parameter to adjust to any of the following major timezones: -12:00, -11:00, -10:00, -09:30, -09:00, -08:00, -07:00, -06:00, -05:00, -04:00, -03:30, -03:00, -02:30, -02:00, -01:00, +00:00, +01:00, +02:00, +03:00, +03:30, +04:00, +04:30, +05:00, +05:30, +05:45, +06:00, +06:30, +07:00, +08:00, +08:45, +09:00, +09:30, +10:00, +10:30, +11:00, +12:00, +12:45, +13:00, +13:45 or +14:00. Note that UTC offsets are treated as-is over the entire timespan, meaning that the offset is not adjusted to account for periods in which daylight savings is honored.}

\item{time_zoom}{This option is only available for timeline modes in HTML format output and enables interactive zooming of the timeline using the browser-based visualization. Set to "yes" to enable and set to "no" or do not include the parameter, to disable. By default, the browser-based timeline display allows interactive examination and export of the timeline data, but does not allow the user to rezoom the display to a more narrow time span. If enabled, the user can click-drag horizontally in the graph to select a specific time period. If the visualization is being displayed directly by itself (it is the "parent" page), it will automatically refresh the page to display the revised time span. If the visualization is being embedded in another page via iframe, it will use postMessage to send the new timespan to the parent page with parameters "startdate" and "enddate" in the format needed by the STARTDATETIME and ENDDATETIME API parameters. The parent page can then use these parameters to rewrite the URLs of any API visualizations embedded in the page and reload each of them. This allows the creation of dashboard-like displays that contain multiple TV API visualizations where the user can zoom the timeline graph at the top and have all of the other displays automatically refresh to narrow their coverage to that revised time frame}
}
\description{
Access data from  \href{https://blog.gdeltproject.org/gdelt-2-0-television-api-debuts/}{GDELT Television V2 API}
}
\examples{
library(gdeltr2)
v2_tv_api(terms = c("Walz", "JD Vance"), formats = c("json"), modes = "TimelineVol")
}
