% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_api.R
\name{v2_tv_trends}
\alias{v2_tv_trends}
\title{Returns Trends from TV API}
\usage{
v2_tv_trends(
  include_national_topics = TRUE,
  return_wide = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{include_national_topics}{if `TRUE` binds national topics}

\item{return_wide}{If `TRUE` returns data in wide form.}

\item{return_message}{If `TRUE` returns messages}
}
\description{
Returns Trends from TV API
}
\examples{
library(gdeltr2)
v2_tv_trends()
v2_tv_trends(return_wide = TRUE)
}
