% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_links.R
\name{v3_social_data}
\alias{v3_social_data}
\title{GDELT Social Media Links}
\usage{
v3_social_data(dates = NULL, return_message = T)
}
\arguments{
\item{dates}{vector of dates}

\item{return_message}{if \code{true} returns message}
}
\description{
Data of articles linking
to social media mentions
}
