% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{visualize_gkg_trelliscope}
\alias{visualize_gkg_trelliscope}
\title{Trelliscope gkg data}
\usage{
visualize_gkg_trelliscope(data, domains = c("wsj.com", "law360",
  "nytimes.com", "dailymail", "recode", "architizer", "globest", "realert",
  "seekingalpha.com", "bloomberg.com", "curbed.com", "commercialobserver.com",
  "globest.com", "realdeal.com", "techcrunch.com", "nytimes.com", "architizer",
  "pehub", "rew-online", "nypost"), random_domains = 0,
  only_pictures = TRUE, extra_columns = NULL, gkg_parse = "names")
}
\arguments{
\item{gkg_parse}{\itemize{
\item 'tone',
\item 'dates',
\item 'events',
\item 'gcam',
\item 'locations',
\item 'names',
\item 'numerics',
\item 'organizations',
\item 'people',
\item 'quotes',
\item 'social',
\item 'themes',
\item 'xml
}}
}
\description{
Trelliscope gkg data
}
