% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{visualize_gkg_tv_trelliscope}
\alias{visualize_gkg_tv_trelliscope}
\title{Trelliscope gkg data}
\usage{
visualize_gkg_tv_trelliscope(data, stations = c("CNN", "Bloomberg",
  "FBC", "BBCNews", "WTTG"), random_stations = 5, only_pictures = TRUE,
  gkg_parse = "people")
}
\arguments{
\item{gkg_parse}{\itemize{
\item 'tone',
\item 'dates',
\item 'events',
\item 'gcam',
\item 'locations',
\item 'names',
\item 'numerics',
\item 'organizations',
\item 'people',
\item 'quotes',
\item 'social',
\item 'themes',
\item 'xml
}}
}
\description{
Trelliscope gkg data
}
