% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_readers.R
\name{widen_gdelt_url_api_parameters}
\alias{widen_gdelt_url_api_parameters}
\title{Widen API URLs}
\usage{
widen_gdelt_url_api_parameters(data, url_gdelt = NULL, widen_variables = NULL)
}
\arguments{
\item{data}{A `tibbile`}

\item{url_gdelt}{Name of GDELT API url variables}

\item{widen_variables}{Vector of variables to unite and widen}
}
\description{
Widen API URLs
}
