% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_ft_api_term}
\alias{get_data_ft_api_term}
\title{Returns data frame for given term from GDELT free text API}
\usage{
get_data_ft_api_term(term = "\\"Brooklyn Nets\\"", domain = NA,
  dedeup_results = T, restrict_to_usa = F, last_minutes = NA,
  max_rows = 1000, only_english = T, return_image_url = T,
  tone_less_than = NA, tone_more_than = NA, search_language = "English",
  source_language = "English", sort_by = "date", return_message = T)
}
\arguments{
\item{term}{any word, can be quoted or not}

\item{domain}{domain name
\code{c(NA,'domain_name')}}

\item{dedeup_results}{return unique results
\code{c(T, F)}}

\item{last_minutes}{how many prior minutes}

\item{max_rows}{number of rows}

\item{only_english}{return only english results
\code{c(T, F)}}

\item{tone_less_than}{tone more than specified number}

\item{tone_more_than}{tone less than specified number}

\item{sort_by}{How do you wish to sort the data
\code{c('date', 'relevence', 'tone.ascending', 'tone.descending')}}

\item{return_message}{\code{c(T, F)}}
}
\value{

}
\description{
Returns data frame for given term from GDELT free text API
}
\examples{
get_data_ft_api_term(term = '"Brooklyn Nets"')
}

