% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gdelt_period_event}
\alias{get_data_gdelt_period_event}
\title{Retreive GDELT data for a given period}
\usage{
get_data_gdelt_period_event(period = 1983, file_directory = NULL,
  folder_name = "gdelt_data", remove_files = T, empty_trash = T,
  return_message = T)
}
\arguments{
\item{period}{the GDELT period}
}
\value{

}
\description{
Retreive GDELT data for a given period
}
\examples{
get_data_gdelt_period_event(period = 1983)
}
