% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_day_summary}
\alias{get_data_gkg_day_summary}
\title{Retrieves gkg summary file for a given day}
\usage{
get_data_gkg_day_summary(date_data = "2016-06-01",
  file_directory = NULL, folder_name = "gdelt_data",
  is_count_file = F, remove_files = T, empty_trash = T,
  return_message = T)
}
\arguments{
\item{date_data}{Date of the data, must be in year, month day form.}

\item{file_directory}{location of where the file is to be saved}

\item{is_count_file}{options
\code{c(TRUE, FALSE)}}

\item{remove_files}{\code{c(TRUE, FALSE)}}

\item{empty_trash}{Empty trash
\code{c(TRUE, FALSE)}}

\item{return_message}{\code{c(TRUE, FALSE)}}
}
\value{

}
\description{
Retrieves gkg summary file for a given day
}
