% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_days_detailed}
\alias{get_data_gkg_days_detailed}
\title{Get dates detailed data from a specified table}
\usage{
get_data_gkg_days_detailed(dates = c("2016-07-19"),
  table_name = c("gkg"), file_directory = NULL,
  folder_name = "gdelt_data", only_most_recent = F, remove_files = T,
  empty_trash = T, return_message = T)
}
\arguments{
\item{table_name}{the name of the table
options \code{c('gkg', 'export', 'mentions'))}}

\item{file_directory}{where are the files to be saved}

\item{remove_files}{Do you want to remove the files
\code{T, F}}

\item{empty_trash}{Do You want to empy the trash
\code{T, F}}

\item{return_message}{Do you want to return a message
\code{T, F}}

\item{date_data}{must be a date in Year - Month - Day format}
}
\description{
Get dates detailed data from a specified table
}
