% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_gkg_tv_day}
\alias{get_data_gkg_tv_day}
\title{Retrieves detailed GKG data for a given day from a specified table}
\usage{
get_data_gkg_tv_day(date_data = "2016-06-01",
  file_directory = "Desktop/temp_gdelt_data", only_most_recent = F,
  remove_files = T, empty_trash = T, return_message = T)
}
\arguments{
\item{date_data}{must be a date in Year - Month - Day format}

\item{table_name}{options \code{c('gkg', 'export', 'mentions'))}}
}
\value{

}
\description{
Retrieves detailed GKG data for a given day from a specified table
}

