% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_free_text_api.R
\name{get_data_locations_instability_api}
\alias{get_data_locations_instability_api}
\title{Returns instability data for given locations}
\usage{
get_data_locations_instability_api(location_ids = c("US", "IS", "TU"),
  variable_names = c("instability", "conflict", "tone", "protest"),
  days_moving_average = NA, time_periods = "daily",
  use_multi_locations = F, return_wide = T, nest_data = F,
  return_message = T)
}
\arguments{
\item{location_ids}{Specify the location IDs}

\item{variable_names}{Specify variables they can include
\code{c('instability', 'conflict', 'protest', 'tone', 'relative mentions')}}

\item{days_moving_average}{Specify day moving average, NA is unsmoothed}

\item{time_periods}{Specified time period
\code{c('daily', '15min')}}

\item{use_multi_locations}{If a country is selected do you want to select all the cities
\code{c(T,F)}}

\item{return_wide}{Return data in wide form}

\item{nest_data}{Return data in nested form
\code{c(T,F)}}

\item{return_message}{Return a message}
}
\description{
Returns instability data for given locations
}

