% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_data_sentiment_ft_api_domains}
\alias{get_data_sentiment_ft_api_domains}
\title{Get V1 sentiment from specified domains}
\usage{
get_data_sentiment_ft_api_domains(domains = c("nytimes.com",
  "washingtonpost.com"), visualization = TRUE, term = NA,
  last_minutes = NA, is_tone = T, tone_less_than = NA,
  tone_more_than = NA, source_language = NA, sort_by = "date",
  dedeup_results = T, nest_data = F, return_message = T)
}
\arguments{
\item{domains}{vector of domain names}

\item{visualization}{if `TRUE` returns a ggplot2 visualization}

\item{term}{specific term}

\item{last_minutes}{how long}

\item{is_tone}{use tone}

\item{tone_less_than}{tone minimum}

\item{tone_more_than}{tone maximum}

\item{source_language}{vector search language}

\item{sort_by}{sort by}

\item{dedeup_results}{if `TRUE` dedupes results}

\item{nest_data}{if `TRUE` returns nested data frame}

\item{return_message}{if `TRUE` returns a messaage}
}
\value{
a `tibble` or a ggplot visualiation
}
\description{
Get V1 sentiment from specified domains
}
\examples{
get_data_sentiment_ft_api_domains(domains = c("foxnews.com", "cnn.com", "washingtonpost.com"), term = "Donald Trump", visualization = T)

}
