% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_data_sentiment_ft_api_terms}
\alias{get_data_sentiment_ft_api_terms}
\title{Get V1 API sentiment for terms}
\usage{
get_data_sentiment_ft_api_terms(terms = c("Zika",
  "\\"Golden State Warriors\\""), visualization = NULL, domain = NA,
  last_minutes = NA, is_tone = T, tone_less_than = NA,
  tone_more_than = NA, source_language = NA, sort_by = "date",
  dedeup_results = T, nest_data = F, return_message = T)
}
\arguments{
\item{terms}{vector of terms}

\item{visualization}{if `TRUE` returns a visualization}

\item{domain}{vector of domain}

\item{last_minutes}{how long}

\item{is_tone}{use tone}

\item{tone_less_than}{tone minimum}

\item{tone_more_than}{tone maximum}

\item{source_language}{vector search language}

\item{sort_by}{sort by}

\item{dedeup_results}{if `TRUE` dedupes results}

\item{nest_data}{if `TRUE` returns nested data frame}

\item{return_message}{if `TRUE` returns a messaage}
}
\value{
a `tibble` or a ggplot visualiation
}
\description{
Get V1 API sentiment for terms
}
\examples{
get_data_sentiment_ft_api_terms(terms = c("Zika", "Brooklyn Nets"), visualization = T)
}
