% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_full_text_api.R
\name{get_data_sentiment_ft_api_terms}
\alias{get_data_sentiment_ft_api_terms}
\title{Returns GDELT full text API sentiment for a given term, can be domain restricted}
\usage{
get_data_sentiment_ft_api_terms(terms = c("Zika",
  "\\"Golden State Warriors\\""), visualization = NULL, domain = NA,
  last_minutes = NA, is_tone = T, tone_less_than = NA,
  tone_more_than = NA, search_language = NA, source_language = NA,
  sort_by = "date", dedeup_results = T, nest_data = F,
  return_message = T)
}
\arguments{
\item{terms}{a vector of terms}

\item{visualization}{\itemize{
\item NULL - returns raw data
\item static returns static visualization
\item interactive returns interactiev visualization
}}

\item{domain}{specificied domain}

\item{last_minutes}{timeperiod to search}

\item{search_language}{langeuage to search}

\item{return_message}{}
}
\description{
Returns GDELT full text API sentiment for a given term, can be domain restricted
}
\examples{
get_data_sentiment_ft_api_terms(terms = c("Zika", '"Golden State Warriors"')) \%>\% View
}
