% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_vgkg_day}
\alias{get_data_vgkg_day}
\title{Get Data CV Day}
\usage{
get_data_vgkg_day(date_data = "2016-06-08", include_translations = F,
  remove_json_column = T, only_most_recent = F, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Get Data CV Day
}

