% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_data_vgkg_url}
\alias{get_data_vgkg_url}
\title{Gets Cloud Vision Data File}
\usage{

  get_data_vgkg_url(url = "http://data.gdeltproject.org/gdeltv2_cloudvision/20160606234500.imagetagsv1.csv.gz",
  file_directory = "Desktop/temp_gdelt_data", remove_files = T,
  empty_trash = T, return_message = T)
}
\description{
Gets Cloud Vision Data File
}

