% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{get_gdelt_url_data}
\alias{get_gdelt_url_data}
\title{Get GDELT url data}
\usage{

  get_gdelt_url_data(url = "http://data.gdeltproject.org/gdeltv2/20160531000000.gkg.csv.zip",
  file_directory = NULL, folder_name = "gdelt_data",
  remove_existing_folder = T, remove_files = T, empty_trash = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get GDELT url data
}
