% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{load_needed_packages}
\alias{load_needed_packages}
\title{Loads needed packages}
\usage{
load_needed_packages(required_packages = c("dplyr"))
}
\arguments{
\item{required_packages}{vector of pckages to lad}
}
\value{
invisible
}
\description{
Loads needed packages
}
