% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_language_types}
\alias{parse_language_types}
\title{Parses XML Language Type}
\usage{
parse_language_types(data, id_date_time = 1)
}
\arguments{
\item{id_date_time}{}
}
\value{

}
\description{
Parses XML Language Type
}

