% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdelt_event_gkg.R
\name{parse_xml_landmarks}
\alias{parse_xml_landmarks}
\title{Parses XML Landmarks}
\usage{
parse_xml_landmarks(data, id_date_time = 886)
}
\arguments{
\item{id_date_time}{}
}
\value{

}
\description{
Parses XML Landmarks
}

